refactor(sessions): move the exit sequence's process reap into its own module
`run.rs` reached 605 lines against the 600-line `PROD-SIZE-1` cap once the exit
sequence gained the non-leader fallback. Pure move: `REAP_BUDGET`,
`kill_process_group_and_reap` and `child_leads_its_own_group` go to
`actor/exit_kill.rs`, whose module doc now carries the reason the kill is aimed
at the group. `run.rs` 605 to 524.
